Adjust notifications: types, entity and persistence
Normalize notification API and persistence handling:

- Change idUsuario parameter types from Long to Integer in controller and service API for consistency.
- Update Notificacion entity: increase "tipo" column length to 50 and remove a duplicate field.
- NotificacionMasivaService: set usuario on Notificacion builder (was usuarioDestino).
- NotificacionServiceImpl: inject EntityManager and use a Hibernate Session.doWork block for privileged DB operations.
- PostulacionServiceImpl: import NotificationRequest and adapt call to enviarNotificacion to pass a NotificationRequest object.
- Remove the marcar-leida endpoint from NotificacionController.
- Change spring.jpa.hibernate.ddl-auto from none to update in application.properties.
- Rename several Flyway migration files (V10->V16, V11->V17, V29->V25).

These changes fix type mismatches, remove duplicate fields, and adjust persistence behavior required by the notification flow.

+++ Backend/backend-sgac/src/main/java/org/uteq/sgacfinal/service/INotificacionService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface INotificacionService {
 
-    NotificacionResponseDTO enviarNotificacion(Long idUsuario, NotificationRequest request);
+    NotificacionResponseDTO enviarNotificacion(Integer idUsuario, NotificationRequest request);
 
     List<NotificacionResponseDTO> listarUltimas10DelUsuarioAutenticado();
